refactor: split Sync tab into focused modules
Split the 1200-line viewer_ui/src/tabs/sync.ts monolith into:
- sync/index.ts — thin orchestrator, data loading, re-exports
- sync/team-sync.ts — coordinator onboarding, invites, join requests
- sync/people.ts — actors, devices, sharing review, legacy claims
- sync/diagnostics.ts — status grid, attempts log, pairing
- sync/helpers.ts — shared helpers, chip editor, module state
- lib/form.ts — validation helpers (markFieldError, friendlyError)

No user-facing behavior changes. All existing exports from the sync
module remain available at the same import path.
