+++ codemem/viewer_static/app.js
@@ -1527,39 +1527,14 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       updateFeedView(true);
     }
   }
-  function shakeField(input) {
-    input.classList.add("sync-shake");
-    input.addEventListener("animationend", () => input.classList.remove("sync-shake"), { once: true });
-  }
-  function markFieldError(input, message) {
-    input.classList.add("sync-field-error");
-    const existing = input.parentElement?.querySelector(".sync-field-hint");
-    if (existing) existing.remove();
-    const hint = document.createElement("div");
-    hint.className = "sync-field-hint";
-    hint.textContent = message;
-    input.insertAdjacentElement("afterend", hint);
-    shakeField(input);
-    input.addEventListener("input", () => clearFieldError(input), { once: true });
-    return false;
-  }
-  function clearFieldError(input) {
-    input.classList.remove("sync-field-error");
-    const hint = input.parentElement?.querySelector(".sync-field-hint");
-    if (hint) hint.remove();
-  }
-  function friendlyError(error, fallback) {
-    if (error instanceof Error) {
-      const msg = error.message;
-      if (msg.includes("fetch") || msg.includes("network") || msg.includes("Failed to fetch")) {
-        return "Network error — check your connection and try again.";
-      }
-      return msg;
-    }
-    return fallback;
-  }
   let adminSetupExpanded = false;
+  function setAdminSetupExpanded(v) {
+    adminSetupExpanded = v;
+  }
   let teamInvitePanelOpen = false;
+  function setTeamInvitePanelOpen(v) {
+    teamInvitePanelOpen = v;
+  }
   const openPeerScopeEditors = /* @__PURE__ */ new Set();
   function redactIpOctets(text) {
     return text.replace(/\b(\d{1,3}\.\d{1,3})\.\d{1,3}\.\d{1,3}\b/g, "$1.#.#");
@@ -1697,24 +1672,10 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       container.appendChild(row);
     });
   }
-  function ensureInvitePanelInAdminSection() {
-    const invitePanel = document.getElementById("syncInvitePanel");
-    const adminSection = document.getElementById("syncAdminSection");
-    if (!invitePanel || !adminSection) return;
-    if (invitePanel.parentElement !== adminSection) adminSection.appendChild(invitePanel);
-  }
-  function ensureJoinPanelInSetupSection() {
-    const joinPanel = document.getElementById("syncJoinPanel");
-    const joinSection = document.getElementById("syncJoinSection");
-    if (!joinPanel || !joinSection) return;
-    if (joinPanel.parentElement !== joinSection) joinSection.appendChild(joinPanel);
-  }
-  function setInviteOutputVisibility() {
-    const syncInviteOutput = document.getElementById("syncInviteOutput");
-    if (!syncInviteOutput) return;
-    const encoded = String(state.lastTeamInvite?.encoded || "").trim();
-    syncInviteOutput.value = encoded;
-    syncInviteOutput.hidden = !encoded;
+  let _renderSyncPeers = () => {
+  };
+  function setRenderSyncPeers(fn) {
+    _renderSyncPeers = fn;
   }
   function renderSyncStatus() {
     const syncStatusGrid = document.getElementById("syncStatusGrid");
@@ -1748,15 +1709,32 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
         `Peers: ${peerCount}`,
         lastSync ? `Last sync: ${formatAgeShort(secondsSince(lastSync))} ago` : "Last sync: never"
       ];
-      if (daemonState === "offline-peers") parts.push("All peers are currently offline; sync will resume automatically");
+      if (daemonState === "offline-peers")
+        parts.push("All peers are currently offline; sync will resume automatically");
       if (daemonDetail && daemonState === "stopped") parts.push(`Detail: ${daemonDetail}`);
       syncMeta.textContent = parts.join(" · ");
     }
-    const diagItems = syncDisabled ? [{ label: "State", value: "Disabled" }, { label: "Mode", value: "Optional" }, { label: "Pending events", value: pending }, { label: "Last sync", value: "n/a" }] : syncNoPeers ? [{ label: "State", value: "No peers" }, { label: "Mode", value: "Idle" }, { label: "Pending events", value: pending }, { label: "Last sync", value: "n/a" }] : [
+    const diagItems = syncDisabled ? [
+      { label: "State", value: "Disabled" },
+      { label: "Mode", value: "Optional" },
+      { label: "Pending events", value: pending },
+      { label: "Last sync", value: "n/a" }
+    ] : syncNoPeers ? [
+      { label: "State", value: "No peers" },
+      { label: "Mode", value: "Idle" },
+      { label: "Pending events", value: pending },
+      { label: "Last sync", value: "n/a" }
+    ] : [
       { label: "State", value: daemonStateLabel },
       { label: "Pending events", value: pending },
-      { label: "Last sync", value: lastSync ? `${formatAgeShort(secondsSince(lastSync))} ago` : "never" },
-      { label: "Last ping", value: lastPing ? `${formatAgeShort(secondsSince(lastPing))} ago` : "never" }
+      {
+        label: "Last sync",
+        value: lastSync ? `${formatAgeShort(secondsSince(lastSync))} ago` : "never"
+      },
+      {
+        label: "Last ping",
+        value: lastPing ? `${formatAgeShort(secondsSince(lastPing))} ago` : "never"
+      }
     ];
     diagItems.forEach((item) => {
       const block = el("div", "stat");
@@ -1768,21 +1746,30 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
     if (!syncDisabled && !syncNoPeers && (syncError || pingError)) {
       const block = el("div", "stat");
       const content = el("div", "stat-content");
-      content.append(el("div", "value", "Errors"), el("div", "label", [syncError, pingError].filter(Boolean).join(" · ")));
+      content.append(
+        el("div", "value", "Errors"),
+        el("div", "label", [syncError, pingError].filter(Boolean).join(" · "))
+      );
       block.appendChild(content);
       syncStatusGrid.appendChild(block);
     }
     if (!syncDisabled && !syncNoPeers && syncPayload?.seconds_since_last) {
       const block = el("div", "stat");
       const content = el("div", "stat-content");
-      content.append(el("div", "value", `${syncPayload.seconds_since_last}s`), el("div", "label", "Since last sync"));
+      content.append(
+        el("div", "value", `${syncPayload.seconds_since_last}s`),
+        el("div", "label", "Since last sync")
+      );
       block.appendChild(content);
       syncStatusGrid.appendChild(block);
     }
     if (!syncDisabled && !syncNoPeers && pingPayload?.seconds_since_last) {
       const block = el("div", "stat");
       const content = el("div", "stat-content");
-      content.append(el("div", "value", `${pingPayload.seconds_since_last}s`), el("div", "label", "Since last ping"));
+      content.append(
+        el("div", "value", `${pingPayload.seconds_since_last}s`),
+        el("div", "label", "Since last ping")
+      );
       block.appendChild(content);
       syncStatusGrid.appendChild(block);
     }
@@ -1793,301 +1780,163 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       actions.push({ label: "Sync daemon is stopped. Start it.", command: "uv run codemem sync start" });
       actions.push({ label: "Then run one immediate sync pass.", command: "uv run codemem sync once" });
     } else if (syncError || pingError || daemonState === "error") {
-      actions.push({ label: "Sync reports errors. Restart now.", command: "uv run codemem sync restart && uv run codemem sync once" });
-      actions.push({ label: "Then run doctor for root cause.", command: "uv run codemem sync doctor" });
+      actions.push({
+        label: "Sync reports errors. Restart now.",
+        command: "uv run codemem sync restart && uv run codemem sync once"
+      });
+      actions.push({
+        label: "Then run doctor for root cause.",
+        command: "uv run codemem sync doctor"
+      });
     } else if (!syncDisabled && !syncNoPeers && pending > 0) {
-      actions.push({ label: "Pending sync work detected. Run one pass now.", command: "uv run codemem sync once" });
+      actions.push({
+        label: "Pending sync work detected. Run one pass now.",
+        command: "uv run codemem sync once"
+      });
     }
     renderActionList(syncActions, actions);
   }
-  function renderSyncPeers() {
-    const syncPeers = document.getElementById("syncPeers");
-    if (!syncPeers) return;
-    syncPeers.textContent = "";
-    const peers = state.lastSyncPeers;
-    if (!Array.isArray(peers) || !peers.length) {
-      syncPeers.appendChild(
+  function renderSyncAttempts() {
+    const syncAttempts = document.getElementById("syncAttempts");
+    if (!syncAttempts) return;
+    syncAttempts.textContent = "";
+    const attempts = state.lastSyncAttempts;
+    if (!Array.isArray(attempts) || !attempts.length) return;
+    attempts.slice(0, 5).forEach((attempt) => {
+      const line = el("div", "diag-line");
+      const left = el("div", "left");
+      left.append(
+        el("div", null, attempt.status || "unknown"),
         el(
           "div",
-          "sync-empty-state",
-          "No devices paired yet. Use the pairing command in Diagnostics to connect another device."
+          "small",
+          isSyncRedactionEnabled() ? redactAddress(attempt.address) : attempt.address || "n/a"
         )
       );
-      return;
-    }
-    peers.forEach((peer) => {
-      const card = el("div", "peer-card");
-      const titleRow = el("div", "peer-title");
-      const peerId = peer.peer_device_id ? String(peer.peer_device_id) : "";
-      const displayName = peer.name || (peerId ? peerId.slice(0, 8) : "unknown");
-      const name = el("strong", null, displayName);
-      if (peerId) name.title = peerId;
-      const peerStatus = peer.status || {};
-      const online = peerStatus.sync_status === "ok" || peerStatus.ping_status === "ok";
-      const badge = el("span", "badge", online ? "Online" : "Offline");
-      badge.style.background = online ? "rgba(31, 111, 92, 0.12)" : "rgba(230, 126, 77, 0.15)";
-      badge.style.color = online ? "var(--accent)" : "var(--accent-warm)";
-      name.append(" ", badge);
-      const actions = el("div", "peer-actions");
-      const primaryAddress = pickPrimaryAddress(peer.addresses);
-      const syncBtn = el("button", null, "Sync now");
-      syncBtn.disabled = !primaryAddress;
-      syncBtn.addEventListener("click", async () => {
-        syncBtn.disabled = true;
-        syncBtn.textContent = "Syncing...";
-        try {
-          await triggerSync(primaryAddress);
-        } catch {
-        }
-        syncBtn.disabled = false;
-        syncBtn.textContent = "Sync now";
-      });
-      actions.appendChild(syncBtn);
-      const toggleScopeBtn = el("button", null, "Edit scope");
-      actions.appendChild(toggleScopeBtn);
-      const peerAddresses = Array.isArray(peer.addresses) ? Array.from(new Set(peer.addresses.filter(Boolean))) : [];
-      const addressLine = peerAddresses.length ? peerAddresses.map((a) => isSyncRedactionEnabled() ? redactAddress(a) : a).join(" · ") : "No addresses";
-      const addressLabel = el("div", "peer-addresses", addressLine);
-      const lastSyncAt = peerStatus.last_sync_at || peerStatus.last_sync_at_utc || "";
-      const lastPingAt = peerStatus.last_ping_at || peerStatus.last_ping_at_utc || "";
-      const meta = el("div", "peer-meta", [
-        lastSyncAt ? `Sync: ${formatTimestamp(lastSyncAt)}` : "Sync: never",
-        lastPingAt ? `Ping: ${formatTimestamp(lastPingAt)}` : "Ping: never"
-      ].join(" · "));
-      const identityMeta = el(
-        "div",
-        "peer-meta",
-        peer.actor_display_name ? `Assigned to ${String(peer.actor_display_name)}${peer.claimed_local_actor ? " · local actor" : ""}` : "Unassigned actor"
-      );
-      const scope = peer.project_scope || {};
-      const includeList = Array.isArray(scope.include) ? scope.include : [];
-      const excludeList = Array.isArray(scope.exclude) ? scope.exclude : [];
-      const effectiveInclude = Array.isArray(scope.effective_include) ? scope.effective_include : [];
-      const effectiveExclude = Array.isArray(scope.effective_exclude) ? scope.effective_exclude : [];
-      const inheritsGlobal = Boolean(scope.inherits_global);
-      const scopePanel = el("div", "peer-scope");
-      const identityRow = el("div", "peer-scope-summary");
-      identityRow.textContent = "Assigned actor";
-      const actorRow = el("div", "peer-actor-row");
-      const actorSelect = document.createElement("select");
-      actorSelect.className = "sync-actor-select";
-      actorSelect.setAttribute("aria-label", `Assigned actor for ${displayName}`);
-      buildActorOptions(String(peer.actor_id || "")).forEach((option) => actorSelect.appendChild(option));
-      const applyActorBtn = el("button", "settings-button", "Save actor");
-      const actorHint = el("div", "peer-scope-effective", assignmentNote(String(peer.actor_id || "")));
-      actorSelect.addEventListener("change", () => {
-        actorHint.textContent = assignmentNote(actorSelect.value);
-      });
-      applyActorBtn.addEventListener("click", async () => {
-        applyActorBtn.disabled = true;
-        actorSelect.disabled = true;
-        applyActorBtn.textContent = "Applying...";
-        try {
-          await assignPeerActor(peerId, actorSelect.value || null);
-          showGlobalNotice(actorSelect.value ? "Device actor updated." : "Device actor cleared.");
-          await loadSyncData();
-        } catch (error) {
-          showGlobalNotice(friendlyError(error, "Failed to update device actor."), "warning");
-          applyActorBtn.textContent = "Retry";
-        } finally {
-          actorSelect.disabled = false;
-          applyActorBtn.disabled = false;
-          if (applyActorBtn.textContent === "Applying...") applyActorBtn.textContent = "Save actor";
-        }
-      });
-      actorRow.append(actorSelect, applyActorBtn);
-      const scopeSummary = el(
-        "div",
-        "peer-scope-summary",
-        inheritsGlobal ? "Using global sync scope" : `Device override · include: ${includeList.join(", ") || "all"} · exclude: ${excludeList.join(", ") || "none"}`
-      );
-      const effectiveSummary = el(
-        "div",
-        "peer-scope-effective",
-        `Effective scope · include: ${effectiveInclude.join(", ") || "all"} · exclude: ${effectiveExclude.join(", ") || "none"}`
-      );
-      const includeEditor = createChipEditor(includeList, "Add included project", "All projects");
-      const excludeEditor = createChipEditor(excludeList, "Add excluded project", "No exclusions");
-      const editorWrap = el("div", "peer-scope-editor-wrap");
-      const scopeEditorOpen = openPeerScopeEditors.has(peerId);
-      editorWrap.hidden = !scopeEditorOpen;
-      const inputRow = el("div", "peer-scope-row");
-      inputRow.append(includeEditor.element, excludeEditor.element);
-      const scopeActions = el("div", "peer-scope-actions");
-      const saveScopeBtn = el("button", "settings-button", "Save scope");
-      const inheritBtn = el("button", "settings-button", "Reset to global scope");
-      saveScopeBtn.addEventListener("click", async () => {
-        saveScopeBtn.disabled = true;
-        saveScopeBtn.textContent = "Saving...";
-        try {
-          await updatePeerScope(peerId, includeEditor.values(), excludeEditor.values());
-          showGlobalNotice("Device sync scope saved.");
-          await loadSyncData();
-        } catch (error) {
-          showGlobalNotice(friendlyError(error, "Failed to save device scope."), "warning");
-          saveScopeBtn.textContent = "Retry save";
-        } finally {
-          saveScopeBtn.disabled = false;
-          if (saveScopeBtn.textContent === "Saving...") saveScopeBtn.textContent = "Save scope";
-        }
-      });
-      inheritBtn.addEventListener("click", async () => {
-        inheritBtn.disabled = true;
-        inheritBtn.textContent = "Resetting...";
-        try {
-          await updatePeerScope(peerId, null, null, true);
-          showGlobalNotice("Device sync scope reset to global defaults.");
-          await loadSyncData();
-        } catch (error) {
-          showGlobalNotice(friendlyError(error, "Failed to reset device scope."), "warning");
-          inheritBtn.textContent = "Retry reset";
-        } finally {
-          inheritBtn.disabled = false;
-          if (inheritBtn.textContent === "Resetting...") inheritBtn.textContent = "Reset to global scope";
-        }
-      });
-      scopeActions.append(saveScopeBtn, inheritBtn);
-      editorWrap.append(inputRow, scopeActions);
-      toggleScopeBtn.textContent = scopeEditorOpen ? "Hide scope editor" : "Edit scope";
-      toggleScopeBtn.addEventListener("click", () => {
-        const nextHidden = !editorWrap.hidden;
-        editorWrap.hidden = nextHidden;
-        if (nextHidden) openPeerScopeEditors.delete(peerId);
-        else openPeerScopeEditors.add(peerId);
-        toggleScopeBtn.textContent = nextHidden ? "Edit scope" : "Hide scope editor";
-      });
-      scopePanel.append(
-        identityRow,
-        identityMeta,
-        actorRow,
-        actorHint,
-        scopeSummary,
-        effectiveSummary,
-        editorWrap
-      );
-      titleRow.append(name, actions);
-      card.append(titleRow, addressLabel, meta, scopePanel);
-      syncPeers.appendChild(card);
+      const right = el("div", "right");
+      const time = attempt.started_at || attempt.started_at_utc || "";
+      right.textContent = time ? formatTimestamp(time) : "";
+      line.append(left, right);
+      syncAttempts.appendChild(line);
     });
   }
-  function renderSyncActors() {
-    const actorList = document.getElementById("syncActorsList");
-    const actorMeta = document.getElementById("syncActorsMeta");
-    if (!actorList) return;
-    actorList.textContent = "";
-    const actors = Array.isArray(state.lastSyncActors) ? state.lastSyncActors : [];
-    if (actorMeta) {
-      actorMeta.textContent = actors.length ? "Create, rename, and merge actors here. Assign each device below. Non-local actors receive memories from allowed projects unless you mark them Only me." : "No named actors yet. Create one here, then assign devices below.";
-    }
-    if (!actors.length) {
-      actorList.appendChild(
-        el(
-          "div",
-          "sync-empty-state",
-          "No actors yet. Create one to represent yourself or a teammate."
-        )
-      );
+  function renderPairing() {
+    const pairingPayloadEl = document.getElementById("pairingPayload");
+    const pairingHint = document.getElementById("pairingHint");
+    if (!pairingPayloadEl) return;
+    const payload = state.pairingPayloadRaw;
+    if (!payload || typeof payload !== "object") {
+      pairingPayloadEl.textContent = "Pairing not available";
+      if (pairingHint) pairingHint.textContent = "Enable sync and retry.";
+      state.pairingCommandRaw = "";
       return;
     }
-    actors.forEach((actor) => {
-      const row = el("div", "actor-row");
-      const details = el("div", "actor-details");
-      const title = el("div", "actor-title");
-      const name = el("strong", null, actorLabel(actor));
-      const count = assignedActorCount(String(actor.actor_id || ""));
-      const badge = el("span", `badge actor-badge${actor.is_local ? " local" : ""}`, actor.is_local ? "Local" : `${count} device${count === 1 ? "" : "s"}`);
-      title.append(name, badge);
-      const note = el(
-        "div",
-        "peer-meta",
-        actor.is_local ? "Used for this device and same-person devices." : `${count} assigned device${count === 1 ? "" : "s"}`
-      );
-      details.append(title, note);
-      const actions = el("div", "actor-actions");
-      if (actor.is_local) {
-        actions.appendChild(el("div", "peer-meta", "Rename in config"));
-      } else {
-        const actorId = String(actor.actor_id || "");
-        const input = document.createElement("input");
-        input.className = "peer-scope-input actor-name-input";
-        input.value = actorLabel(actor);
-        input.setAttribute("aria-label", `Rename ${actorLabel(actor)}`);
-        const renameBtn = el("button", "settings-button", "Rename");
-        renameBtn.addEventListener("click", async () => {
-          const nextName = input.value.trim();
-          if (!nextName) return;
-          renameBtn.disabled = true;
-          input.disabled = true;
-          renameBtn.textContent = "Saving...";
-          try {
-            await renameActor(actorId, nextName);
-            await loadSyncData();
-          } catch {
-            renameBtn.textContent = "Retry rename";
-          } finally {
-            renameBtn.disabled = false;
-            input.disabled = false;
-            if (renameBtn.textContent === "Saving...") renameBtn.textContent = "Rename";
-          }
-        });
-        const mergeTargets = mergeTargetActors(actorId);
-        const mergeControls = el("div", "actor-merge-controls");
-        const mergeSelect = document.createElement("select");
-        mergeSelect.className = "sync-actor-select actor-merge-select";
-        mergeSelect.setAttribute("aria-label", `Merge ${actorLabel(actor)} into another actor`);
-        const placeholder = document.createElement("option");
-        placeholder.value = "";
-        placeholder.textContent = "Merge into actor";
-        placeholder.selected = true;
-        mergeSelect.appendChild(placeholder);
-        mergeTargets.forEach((target) => {
-          const option = document.createElement("option");
-          option.value = String(target.actor_id || "");
-          option.textContent = target.is_local ? `${actorLabel(target)} (local)` : actorLabel(target);
-          mergeSelect.appendChild(option);
-        });
-        const mergeBtn = el("button", "settings-button", "Merge into selected actor");
-        mergeBtn.disabled = mergeTargets.length === 0;
-        const mergeNote = el(
-          "div",
-          "peer-meta actor-merge-note",
-          mergeTargets.length ? actorMergeNote("", actorId) : "No merge targets yet. Create another actor or use the local actor."
-        );
-        mergeSelect.addEventListener("change", () => {
-          mergeNote.textContent = actorMergeNote(mergeSelect.value, actorId);
-        });
-        mergeBtn.addEventListener("click", async () => {
-          if (!mergeSelect.value) return;
-          const target = mergeTargets.find((candidate) => String(candidate.actor_id || "") === mergeSelect.value);
-          if (!window.confirm(`Merge ${actorLabel(actor)} into ${actorLabel(target)}? Assigned devices move now, but older memories keep their current stamped provenance for now.`)) {
-            return;
-          }
-          mergeBtn.disabled = true;
-          mergeSelect.disabled = true;
-          input.disabled = true;
-          renameBtn.disabled = true;
-          mergeBtn.textContent = "Merging...";
-          try {
-            await mergeActor(mergeSelect.value, actorId);
-            showGlobalNotice("Actor merged. Assigned devices moved to the selected actor.");
-            await loadSyncData();
-          } catch (error) {
-            showGlobalNotice(friendlyError(error, "Failed to merge actor."), "warning");
-            mergeBtn.textContent = "Retry merge";
-          } finally {
-            mergeBtn.disabled = mergeTargets.length === 0;
-            mergeSelect.disabled = false;
-            input.disabled = false;
-            renameBtn.disabled = false;
-            if (mergeBtn.textContent === "Merging...") mergeBtn.textContent = "Merge into selected actor";
-          }
-        });
-        mergeControls.append(mergeSelect, mergeBtn);
-        actions.append(input, renameBtn, mergeControls, mergeNote);
+    if (payload.redacted) {
+      pairingPayloadEl.textContent = "Pairing payload hidden";
+      if (pairingHint)
+        pairingHint.textContent = "Diagnostics are required to view the pairing payload.";
+      state.pairingCommandRaw = "";
+      return;
+    }
+    const safePayload = {
+      ...payload,
+      addresses: Array.isArray(payload.addresses) ? payload.addresses : []
+    };
+    const compact = JSON.stringify(safePayload);
+    const b64 = btoa(compact);
+    const command = `echo '${b64}' | base64 -d | codemem sync pair --accept-file -`;
+    pairingPayloadEl.textContent = command;
+    state.pairingCommandRaw = command;
+    if (pairingHint) {
+      pairingHint.textContent = "Copy this command and run it on the other device. Use --include/--exclude to control which projects sync.";
+    }
+  }
+  function initDiagnosticsEvents(refreshCallback) {
+    const syncPairingToggle = document.getElementById(
+      "syncPairingToggle"
+    );
+    const syncRedact = document.getElementById("syncRedact");
+    const pairingCopy = document.getElementById("pairingCopy");
+    const syncPairing = document.getElementById("syncPairing");
+    if (syncPairing) syncPairing.hidden = !state.syncPairingOpen;
+    if (syncPairingToggle)
+      syncPairingToggle.textContent = state.syncPairingOpen ? "Hide pairing" : "Show pairing";
+    if (syncRedact) syncRedact.checked = isSyncRedactionEnabled();
+    syncPairingToggle?.addEventListener("click", () => {
+      const next = !state.syncPairingOpen;
+      setSyncPairingOpen(next);
+      if (syncPairing) syncPairing.hidden = !next;
+      if (syncPairingToggle)
+        syncPairingToggle.textContent = next ? "Hide pairing" : "Show pairing";
+      if (next) {
+        const pairingPayloadEl = document.getElementById("pairingPayload");
+        const pairingHint = document.getElementById("pairingHint");
+        if (pairingPayloadEl) pairingPayloadEl.textContent = "Loading…";
+        if (pairingHint) pairingHint.textContent = "Fetching pairing payload…";
       }
-      row.append(details, actions);
-      actorList.appendChild(row);
+      refreshCallback();
     });
+    syncRedact?.addEventListener("change", () => {
+      setSyncRedactionEnabled(Boolean(syncRedact.checked));
+      renderSyncStatus();
+      _renderSyncPeers();
+      renderSyncAttempts();
+      renderPairing();
+    });
+    pairingCopy?.addEventListener("click", async () => {
+      const text = state.pairingCommandRaw || document.getElementById("pairingPayload")?.textContent || "";
+      if (text && pairingCopy) await copyToClipboard(text, pairingCopy);
+    });
+  }
+  function shakeField(input) {
+    input.classList.add("sync-shake");
+    input.addEventListener("animationend", () => input.classList.remove("sync-shake"), { once: true });
+  }
+  function markFieldError(input, message) {
+    input.classList.add("sync-field-error");
+    const existing = input.parentElement?.querySelector(".sync-field-hint");
+    if (existing) existing.remove();
+    const hint = document.createElement("div");
+    hint.className = "sync-field-hint";
+    hint.textContent = message;
+    input.insertAdjacentElement("afterend", hint);
+    shakeField(input);
+    input.addEventListener("input", () => clearFieldError(input), { once: true });
+    return false;
+  }
+  function clearFieldError(input) {
+    input.classList.remove("sync-field-error");
+    const hint = input.parentElement?.querySelector(".sync-field-hint");
+    if (hint) hint.remove();
+  }
+  function friendlyError(error, fallback) {
+    if (error instanceof Error) {
+      const msg = error.message;
+      if (msg.includes("fetch") || msg.includes("network") || msg.includes("Failed to fetch")) {
+        return "Network error — check your connection and try again.";
+      }
+      return msg;
+    }
+    return fallback;
+  }
+  function ensureInvitePanelInAdminSection() {
+    const invitePanel = document.getElementById("syncInvitePanel");
+    const adminSection = document.getElementById("syncAdminSection");
+    if (!invitePanel || !adminSection) return;
+    if (invitePanel.parentElement !== adminSection) adminSection.appendChild(invitePanel);
+  }
+  function ensureJoinPanelInSetupSection() {
+    const joinPanel = document.getElementById("syncJoinPanel");
+    const joinSection = document.getElementById("syncJoinSection");
+    if (!joinPanel || !joinSection) return;
+    if (joinPanel.parentElement !== joinSection) joinSection.appendChild(joinPanel);
+  }
+  function setInviteOutputVisibility() {
+    const syncInviteOutput = document.getElementById("syncInviteOutput");
+    if (!syncInviteOutput) return;
+    const encoded = String(state.lastTeamInvite?.encoded || "").trim();
+    syncInviteOutput.value = encoded;
+    syncInviteOutput.hidden = !encoded;
   }
   function openFeedSharingReview() {
     setFeedScopeFilter("mine");
@@ -2114,7 +1963,11 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       const title = el("div", "actor-title");
       title.append(
         el("strong", null, String(item.peer_name || item.peer_device_id || "Device")),
-        el("span", "badge actor-badge", `actor: ${String(item.actor_display_name || item.actor_id || "unknown")}`)
+        el(
+          "span",
+          "badge actor-badge",
+          `actor: ${String(item.actor_display_name || item.actor_id || "unknown")}`
+        )
       );
       const note = el(
         "div",
@@ -2124,13 +1977,16 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       details.append(title, note);
       const actions = el("div", "actor-actions");
       const reviewBtn = el("button", "settings-button", "Review my memories in Feed");
-      reviewBtn.addEventListener("click", () => {
-        openFeedSharingReview();
-      });
+      reviewBtn.addEventListener("click", () => openFeedSharingReview());
       actions.appendChild(reviewBtn);
       row.append(details, actions);
       list.appendChild(row);
     });
+  }
+  let _loadSyncData$1 = async () => {
+  };
+  function setLoadSyncData$1(fn) {
+    _loadSyncData$1 = fn;
   }
   function renderTeamSync() {
     const meta = document.getElementById("syncTeamMeta");
@@ -2188,10 +2044,14 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
     const inviteToggleRow = el("div", "sync-action");
     const inviteToggleText = el("div", "sync-action-text");
     inviteToggleText.textContent = "Generate an invite to add another teammate to this team.";
-    const inviteToggleBtn = el("button", "settings-button", "Invite a teammate");
+    const inviteToggleBtn = el(
+      "button",
+      "settings-button",
+      "Invite a teammate"
+    );
     inviteToggleBtn.addEventListener("click", () => {
       if (!invitePanel) return;
-      teamInvitePanelOpen = !teamInvitePanelOpen;
+      setTeamInvitePanelOpen(!teamInvitePanelOpen);
       if (invitePanel.parentElement !== actions) actions.appendChild(invitePanel);
       invitePanel.hidden = !teamInvitePanelOpen;
       inviteToggleBtn.textContent = teamInvitePanelOpen ? "Hide invite form" : "Invite a teammate";
@@ -2216,7 +2076,11 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       const textWrap = el("div", "sync-action-text");
       textWrap.textContent = "This device is not connected to the team yet.";
       textWrap.appendChild(
-        el("span", "sync-action-command", "Import a team invite or ask your admin to enroll this device")
+        el(
+          "span",
+          "sync-action-command",
+          "Import a team invite or ask your admin to enroll this device"
+        )
       );
       actions.appendChild(row);
       row.appendChild(textWrap);
@@ -2225,15 +2089,24 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       const row = el("div", "sync-action");
       const textWrap = el("div", "sync-action-text");
       textWrap.textContent = "No devices are paired yet.";
-      textWrap.appendChild(el("span", "sync-action-command", "uv run codemem sync pair --payload-only"));
+      textWrap.appendChild(
+        el("span", "sync-action-command", "uv run codemem sync pair --payload-only")
+      );
       const btn = el("button", "settings-button sync-action-copy", "Copy");
-      btn.addEventListener("click", () => copyToClipboard("uv run codemem sync pair --payload-only", btn));
+      btn.addEventListener(
+        "click",
+        () => copyToClipboard("uv run codemem sync pair --payload-only", btn)
+      );
       row.append(textWrap, btn);
       actions.appendChild(row);
     }
     const pending = Array.isArray(state.lastSyncJoinRequests) ? state.lastSyncJoinRequests : [];
     if (joinRequests && pending.length) {
-      const title = el("div", "peer-meta", `${pending.length} pending join request${pending.length === 1 ? "" : "s"}`);
+      const title = el(
+        "div",
+        "peer-meta",
+        `${pending.length} pending join request${pending.length === 1 ? "" : "s"}`
+      );
       joinRequests.appendChild(title);
       pending.forEach((request) => {
         const row = el("div", "actor-row");
@@ -2253,7 +2126,7 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
           try {
             await reviewJoinRequest(String(request.request_id || ""), "approve");
             showGlobalNotice(`Approved ${name}. They can now sync with the team.`);
-            await loadSyncData();
+            await _loadSyncData$1();
           } catch (error) {
             showGlobalNotice(friendlyError(error, "Failed to approve join request."), "warning");
             approveBtn.textContent = "Retry";
@@ -2263,14 +2136,17 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
           }
         });
         denyBtn.addEventListener("click", async () => {
-          if (!window.confirm(`Deny join request from ${name}? They will need a new invite to try again.`)) return;
+          if (!window.confirm(
+            `Deny join request from ${name}? They will need a new invite to try again.`
+          ))
+            return;
           approveBtn.disabled = true;
           denyBtn.disabled = true;
           denyBtn.textContent = "Denying…";
           try {
             await reviewJoinRequest(String(request.request_id || ""), "deny");
             showGlobalNotice(`Denied join request from ${name}.`);
-            await loadSyncData();
+            await _loadSyncData$1();
           } catch (error) {
             showGlobalNotice(friendlyError(error, "Failed to deny join request."), "warning");
             denyBtn.textContent = "Retry deny";
@@ -2287,207 +2163,30 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       joinRequests.hidden = true;
     }
   }
-  function renderLegacyDeviceClaims() {
-    const panel = document.getElementById("syncLegacyClaims");
-    const select = document.getElementById("syncLegacyDeviceSelect");
-    const button = document.getElementById("syncLegacyClaimButton");
-    const meta = document.getElementById("syncLegacyClaimsMeta");
-    if (!panel || !select || !button || !meta) return;
-    const devices = Array.isArray(state.lastSyncLegacyDevices) ? state.lastSyncLegacyDevices : [];
-    select.textContent = "";
-    meta.textContent = "";
-    if (!devices.length) {
-      panel.hidden = true;
-      return;
-    }
-    panel.hidden = false;
-    devices.forEach((device, index) => {
-      const option = document.createElement("option");
-      const deviceId = String(device.origin_device_id || "").trim();
-      if (!deviceId) return;
-      const count = Number(device.memory_count || 0);
-      const lastSeen = String(device.last_seen_at || "").trim();
-      option.value = deviceId;
-      option.textContent = count > 0 ? `${deviceId} (${count} memories)` : deviceId;
-      if (index === 0) option.selected = true;
-      select.appendChild(option);
-      if (!meta.textContent && lastSeen) {
-        meta.textContent = `Detected from older synced memories. Latest memory: ${formatTimestamp(lastSeen)}`;
-      }
-    });
-    if (!meta.textContent) {
-      meta.textContent = "Detected from older synced memories not yet attached to a current device.";
-    }
-  }
-  function renderSyncAttempts() {
-    const syncAttempts = document.getElementById("syncAttempts");
-    if (!syncAttempts) return;
-    syncAttempts.textContent = "";
-    const attempts = state.lastSyncAttempts;
-    if (!Array.isArray(attempts) || !attempts.length) return;
-    attempts.slice(0, 5).forEach((attempt) => {
-      const line = el("div", "diag-line");
-      const left = el("div", "left");
-      left.append(
-        el("div", null, attempt.status || "unknown"),
-        el("div", "small", isSyncRedactionEnabled() ? redactAddress(attempt.address) : attempt.address || "n/a")
-      );
-      const right = el("div", "right");
-      const time = attempt.started_at || attempt.started_at_utc || "";
-      right.textContent = time ? formatTimestamp(time) : "";
-      line.append(left, right);
-      syncAttempts.appendChild(line);
-    });
-  }
-  function renderPairing() {
-    const pairingPayloadEl = document.getElementById("pairingPayload");
-    const pairingHint = document.getElementById("pairingHint");
-    if (!pairingPayloadEl) return;
-    const payload = state.pairingPayloadRaw;
-    if (!payload || typeof payload !== "object") {
-      pairingPayloadEl.textContent = "Pairing not available";
-      if (pairingHint) pairingHint.textContent = "Enable sync and retry.";
-      state.pairingCommandRaw = "";
-      return;
-    }
-    if (payload.redacted) {
-      pairingPayloadEl.textContent = "Pairing payload hidden";
-      if (pairingHint) pairingHint.textContent = "Diagnostics are required to view the pairing payload.";
-      state.pairingCommandRaw = "";
-      return;
-    }
-    const safePayload = { ...payload, addresses: Array.isArray(payload.addresses) ? payload.addresses : [] };
-    const compact = JSON.stringify(safePayload);
-    const b64 = btoa(compact);
-    const command = `echo '${b64}' | base64 -d | codemem sync pair --accept-file -`;
-    pairingPayloadEl.textContent = command;
-    state.pairingCommandRaw = command;
-    if (pairingHint) {
-      pairingHint.textContent = "Copy this command and run it on the other device. Use --include/--exclude to control which projects sync.";
-    }
-  }
-  async function loadSyncData() {
-    try {
-      const payload = await loadSyncStatus(true, state.currentProject || "");
-      let actorsPayload = null;
-      let actorLoadError = false;
-      try {
-        actorsPayload = await loadSyncActors();
-      } catch {
-        actorLoadError = true;
-      }
-      const statusPayload = payload.status && typeof payload.status === "object" ? payload.status : null;
-      if (statusPayload) state.lastSyncStatus = statusPayload;
-      state.lastSyncActors = Array.isArray(actorsPayload?.items) ? actorsPayload.items : [];
-      state.lastSyncPeers = payload.peers || [];
-      state.lastSyncSharingReview = payload.sharing_review || [];
-      state.lastSyncCoordinator = payload.coordinator || null;
-      state.lastSyncJoinRequests = payload.join_requests || [];
-      state.lastSyncAttempts = payload.attempts || [];
-      state.lastSyncLegacyDevices = payload.legacy_devices || [];
-      renderSyncStatus();
-      renderTeamSync();
-      renderSyncActors();
-      renderSyncSharingReview();
-      renderSyncPeers();
-      renderLegacyDeviceClaims();
-      renderSyncAttempts();
-      renderHealthOverview();
-      if (actorLoadError) {
-        const actorMeta = document.getElementById("syncActorsMeta");
-        if (actorMeta) actorMeta.textContent = "Actor controls are temporarily unavailable. Peer status and sync health still loaded.";
-      }
-    } catch {
-      const syncMeta = document.getElementById("syncMeta");
-      if (syncMeta) syncMeta.textContent = "Sync unavailable";
-    }
-  }
-  async function loadPairingData() {
-    try {
-      const payload = await loadPairing();
-      state.pairingPayloadRaw = payload || null;
-      renderPairing();
-    } catch {
-      renderPairing();
-    }
-  }
-  function initSyncTab(refreshCallback) {
-    const syncPairingToggle = document.getElementById("syncPairingToggle");
+  function initTeamSyncEvents(refreshCallback, loadSyncData2) {
     const syncNowButton = document.getElementById("syncNowButton");
-    const syncRedact = document.getElementById("syncRedact");
-    const syncActorCreateButton = document.getElementById("syncActorCreateButton");
-    const syncActorCreateInput = document.getElementById("syncActorCreateInput");
-    const syncCreateInviteButton = document.getElementById("syncCreateInviteButton");
+    const syncToggleAdmin = document.getElementById("syncToggleAdmin");
+    const syncInvitePanel = document.getElementById("syncInvitePanel");
+    const syncCreateInviteButton = document.getElementById(
+      "syncCreateInviteButton"
+    );
     const syncInviteGroup = document.getElementById("syncInviteGroup");
     const syncInvitePolicy = document.getElementById("syncInvitePolicy");
     const syncInviteTtl = document.getElementById("syncInviteTtl");
-    const syncInviteOutput = document.getElementById("syncInviteOutput");
-    const syncToggleAdmin = document.getElementById("syncToggleAdmin");
-    const syncInvitePanel = document.getElementById("syncInvitePanel");
+    const syncInviteOutput = document.getElementById(
+      "syncInviteOutput"
+    );
     const syncJoinButton = document.getElementById("syncJoinButton");
     const syncJoinInvite = document.getElementById("syncJoinInvite");
-    const syncLegacyClaimButton = document.getElementById("syncLegacyClaimButton");
-    const syncLegacyDeviceSelect = document.getElementById("syncLegacyDeviceSelect");
-    const pairingCopy = document.getElementById("pairingCopy");
-    const syncPairing = document.getElementById("syncPairing");
-    if (syncPairing) syncPairing.hidden = !state.syncPairingOpen;
-    if (syncPairingToggle) syncPairingToggle.textContent = state.syncPairingOpen ? "Hide pairing" : "Show pairing";
-    if (syncRedact) syncRedact.checked = isSyncRedactionEnabled();
-    syncPairingToggle?.addEventListener("click", () => {
-      const next = !state.syncPairingOpen;
-      setSyncPairingOpen(next);
-      if (syncPairing) syncPairing.hidden = !next;
-      if (syncPairingToggle) syncPairingToggle.textContent = next ? "Hide pairing" : "Show pairing";
-      if (next) {
-        const pairingPayloadEl = document.getElementById("pairingPayload");
-        const pairingHint = document.getElementById("pairingHint");
-        if (pairingPayloadEl) pairingPayloadEl.textContent = "Loading…";
-        if (pairingHint) pairingHint.textContent = "Fetching pairing payload…";
-      }
-      refreshCallback();
-    });
-    syncRedact?.addEventListener("change", () => {
-      setSyncRedactionEnabled(Boolean(syncRedact.checked));
-      renderSyncStatus();
-      renderSyncPeers();
-      renderSyncAttempts();
-      renderPairing();
-    });
-    syncActorCreateButton?.addEventListener("click", async () => {
-      if (!syncActorCreateButton || !syncActorCreateInput) return;
-      const displayName = String(syncActorCreateInput.value || "").trim();
-      if (!displayName) {
-        markFieldError(syncActorCreateInput, "Enter a name for the actor.");
-        return;
-      }
-      clearFieldError(syncActorCreateInput);
-      syncActorCreateButton.disabled = true;
-      syncActorCreateInput.disabled = true;
-      syncActorCreateButton.textContent = "Creating...";
-      try {
-        await createActor(displayName);
-        showGlobalNotice("Actor created.");
-        syncActorCreateInput.value = "";
-        await loadSyncData();
-      } catch (error) {
-        showGlobalNotice(friendlyError(error, "Failed to create actor."), "warning");
-        syncActorCreateButton.textContent = "Retry";
-        syncActorCreateButton.disabled = false;
-        syncActorCreateInput.disabled = false;
-        return;
-      }
-      syncActorCreateButton.textContent = "Create actor";
-      syncActorCreateButton.disabled = false;
-      syncActorCreateInput.disabled = false;
-    });
     syncToggleAdmin?.addEventListener("click", () => {
       if (!syncInvitePanel) return;
-      adminSetupExpanded = !adminSetupExpanded;
+      setAdminSetupExpanded(!adminSetupExpanded);
       syncInvitePanel.hidden = !adminSetupExpanded;
       syncToggleAdmin.textContent = adminSetupExpanded ? "Hide team setup" : "Set up a new team instead…";
     });
     syncCreateInviteButton?.addEventListener("click", async () => {
-      if (!syncCreateInviteButton || !syncInviteGroup || !syncInvitePolicy || !syncInviteTtl || !syncInviteOutput) return;
+      if (!syncCreateInviteButton || !syncInviteGroup || !syncInvitePolicy || !syncInviteTtl || !syncInviteOutput)
+        return;
       const groupName = syncInviteGroup.value.trim();
       const ttlValue = Number(syncInviteTtl.value);
       let valid = true;
@@ -2536,9 +2235,11 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       try {
         const result = await importCoordinatorInvite(inviteValue);
         state.lastTeamJoin = result;
-        showGlobalNotice(result.status === "pending" ? "Join request submitted — waiting for admin approval." : "Joined team successfully.");
+        showGlobalNotice(
+          result.status === "pending" ? "Join request submitted — waiting for admin approval." : "Joined team successfully."
+        );
         syncJoinInvite.value = "";
-        await loadSyncData();
+        await loadSyncData2();
       } catch (error) {
         showGlobalNotice(friendlyError(error, "Failed to import invite."), "warning");
       } finally {
@@ -2546,30 +2247,10 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
         syncJoinButton.textContent = "Join team";
       }
     });
-    syncLegacyClaimButton?.addEventListener("click", async () => {
-      const originDeviceId = String(syncLegacyDeviceSelect?.value || "").trim();
-      if (!originDeviceId || !syncLegacyClaimButton) return;
-      if (!window.confirm(`Attach old device history from ${originDeviceId} to your local actor? This updates legacy provenance for that device.`)) return;
-      syncLegacyClaimButton.disabled = true;
-      const originalText = syncLegacyClaimButton.textContent || "Attach device history";
-      syncLegacyClaimButton.textContent = "Attaching...";
-      try {
-        await claimLegacyDeviceIdentity(originDeviceId);
-        showGlobalNotice("Old device history attached to your local actor.");
-        await loadSyncData();
-      } catch (error) {
-        showGlobalNotice(friendlyError(error, "Failed to attach old device history."), "warning");
-        syncLegacyClaimButton.textContent = "Retry";
-        syncLegacyClaimButton.disabled = false;
-        return;
-      }
-      syncLegacyClaimButton.textContent = originalText;
-      syncLegacyClaimButton.disabled = false;
-    });
     syncNowButton?.addEventListener("click", async () => {
       if (!syncNowButton) return;
       syncNowButton.disabled = true;
-      syncNowButton.textContent = "Syncing...";
+      syncNowButton.textContent = "Syncing…";
       try {
         await triggerSync();
         showGlobalNotice("Sync pass started.");
@@ -2580,10 +2261,467 @@ Global: ${Number(totalsGlobal.tokens_saved || 0).toLocaleString()} saved` : "";
       syncNowButton.textContent = "Sync now";
       refreshCallback();
     });
-    pairingCopy?.addEventListener("click", async () => {
-      const text = state.pairingCommandRaw || document.getElementById("pairingPayload")?.textContent || "";
-      if (text && pairingCopy) await copyToClipboard(text, pairingCopy);
+  }
+  let _loadSyncData = async () => {
+  };
+  function setLoadSyncData(fn) {
+    _loadSyncData = fn;
+  }
+  function renderSyncActors() {
+    const actorList = document.getElementById("syncActorsList");
+    const actorMeta = document.getElementById("syncActorsMeta");
+    if (!actorList) return;
+    actorList.textContent = "";
+    const actors = Array.isArray(state.lastSyncActors) ? state.lastSyncActors : [];
+    if (actorMeta) {
+      actorMeta.textContent = actors.length ? "Create, rename, and merge actors here. Assign each device below. Non-local actors receive memories from allowed projects unless you mark them Only me." : "No named actors yet. Create one here, then assign devices below.";
+    }
+    if (!actors.length) {
+      actorList.appendChild(
+        el("div", "sync-empty-state", "No actors yet. Create one to represent yourself or a teammate.")
+      );
+      return;
+    }
+    actors.forEach((actor) => {
+      const row = el("div", "actor-row");
+      const details = el("div", "actor-details");
+      const title = el("div", "actor-title");
+      const name = el("strong", null, actorLabel(actor));
+      const count = assignedActorCount(String(actor.actor_id || ""));
+      const badge = el(
+        "span",
+        `badge actor-badge${actor.is_local ? " local" : ""}`,
+        actor.is_local ? "Local" : `${count} device${count === 1 ? "" : "s"}`
+      );
+      title.append(name, badge);
+      const note = el(
+        "div",
+        "peer-meta",
+        actor.is_local ? "Used for this device and same-person devices." : `${count} assigned device${count === 1 ? "" : "s"}`
+      );
+      details.append(title, note);
+      const actions = el("div", "actor-actions");
+      if (actor.is_local) {
+        actions.appendChild(el("div", "peer-meta", "Rename in config"));
+      } else {
+        const actorId = String(actor.actor_id || "");
+        const input = document.createElement("input");
+        input.className = "peer-scope-input actor-name-input";
+        input.value = actorLabel(actor);
+        input.setAttribute("aria-label", `Rename ${actorLabel(actor)}`);
+        const renameBtn = el("button", "settings-button", "Rename");
+        renameBtn.addEventListener("click", async () => {
+          const nextName = input.value.trim();
+          if (!nextName) return;
+          renameBtn.disabled = true;
+          input.disabled = true;
+          renameBtn.textContent = "Saving…";
+          try {
+            await renameActor(actorId, nextName);
+            await _loadSyncData();
+          } catch {
+            renameBtn.textContent = "Retry rename";
+          } finally {
+            renameBtn.disabled = false;
+            input.disabled = false;
+            if (renameBtn.textContent === "Saving…") renameBtn.textContent = "Rename";
+          }
+        });
+        const mergeTargets = mergeTargetActors(actorId);
+        const mergeControls = el("div", "actor-merge-controls");
+        const mergeSelect = document.createElement("select");
+        mergeSelect.className = "sync-actor-select actor-merge-select";
+        mergeSelect.setAttribute("aria-label", `Merge ${actorLabel(actor)} into another actor`);
+        const placeholder = document.createElement("option");
+        placeholder.value = "";
+        placeholder.textContent = "Merge into actor";
+        placeholder.selected = true;
+        mergeSelect.appendChild(placeholder);
+        mergeTargets.forEach((target) => {
+          const option = document.createElement("option");
+          option.value = String(target.actor_id || "");
+          option.textContent = target.is_local ? `${actorLabel(target)} (local)` : actorLabel(target);
+          mergeSelect.appendChild(option);
+        });
+        const mergeBtn = el(
+          "button",
+          "settings-button",
+          "Merge into selected actor"
+        );
+        mergeBtn.disabled = mergeTargets.length === 0;
+        const mergeNote = el(
+          "div",
+          "peer-meta actor-merge-note",
+          mergeTargets.length ? actorMergeNote("", actorId) : "No merge targets yet. Create another actor or use the local actor."
+        );
+        mergeSelect.addEventListener("change", () => {
+          mergeNote.textContent = actorMergeNote(mergeSelect.value, actorId);
+        });
+        mergeBtn.addEventListener("click", async () => {
+          if (!mergeSelect.value) return;
+          const target = mergeTargets.find(
+            (candidate) => String(candidate.actor_id || "") === mergeSelect.value
+          );
+          if (!window.confirm(
+            `Merge ${actorLabel(actor)} into ${actorLabel(target)}? Assigned devices move now, but older memories keep their current stamped provenance for now.`
+          )) {
+            return;
+          }
+          mergeBtn.disabled = true;
+          mergeSelect.disabled = true;
+          input.disabled = true;
+          renameBtn.disabled = true;
+          mergeBtn.textContent = "Merging…";
+          try {
+            await mergeActor(mergeSelect.value, actorId);
+            showGlobalNotice("Actor merged. Assigned devices moved to the selected actor.");
+            await _loadSyncData();
+          } catch (error) {
+            showGlobalNotice(friendlyError(error, "Failed to merge actor."), "warning");
+            mergeBtn.textContent = "Retry merge";
+          } finally {
+            mergeBtn.disabled = mergeTargets.length === 0;
+            mergeSelect.disabled = false;
+            input.disabled = false;
+            renameBtn.disabled = false;
+            if (mergeBtn.textContent === "Merging…")
+              mergeBtn.textContent = "Merge into selected actor";
+          }
+        });
+        mergeControls.append(mergeSelect, mergeBtn);
+        actions.append(input, renameBtn, mergeControls, mergeNote);
+      }
+      row.append(details, actions);
+      actorList.appendChild(row);
     });
+  }
+  function renderSyncPeers() {
+    const syncPeers = document.getElementById("syncPeers");
+    if (!syncPeers) return;
+    syncPeers.textContent = "";
+    const peers = state.lastSyncPeers;
+    if (!Array.isArray(peers) || !peers.length) {
+      syncPeers.appendChild(
+        el(
+          "div",
+          "sync-empty-state",
+          "No devices paired yet. Use the pairing command in Diagnostics to connect another device."
+        )
+      );
+      return;
+    }
+    peers.forEach((peer) => {
+      const card = el("div", "peer-card");
+      const titleRow = el("div", "peer-title");
+      const peerId = peer.peer_device_id ? String(peer.peer_device_id) : "";
+      const displayName = peer.name || (peerId ? peerId.slice(0, 8) : "unknown");
+      const name = el("strong", null, displayName);
+      if (peerId) name.title = peerId;
+      const peerStatus = peer.status || {};
+      const online = peerStatus.sync_status === "ok" || peerStatus.ping_status === "ok";
+      const badge = el("span", "badge", online ? "Online" : "Offline");
+      badge.style.background = online ? "rgba(31, 111, 92, 0.12)" : "rgba(230, 126, 77, 0.15)";
+      badge.style.color = online ? "var(--accent)" : "var(--accent-warm)";
+      name.append(" ", badge);
+      const actions = el("div", "peer-actions");
+      const primaryAddress = pickPrimaryAddress(peer.addresses);
+      const syncBtn = el("button", null, "Sync now");
+      syncBtn.disabled = !primaryAddress;
+      syncBtn.addEventListener("click", async () => {
+        syncBtn.disabled = true;
+        syncBtn.textContent = "Syncing…";
+        try {
+          await triggerSync(primaryAddress);
+        } catch {
+        }
+        syncBtn.disabled = false;
+        syncBtn.textContent = "Sync now";
+      });
+      actions.appendChild(syncBtn);
+      const toggleScopeBtn = el("button", null, "Edit scope");
+      actions.appendChild(toggleScopeBtn);
+      const peerAddresses = Array.isArray(peer.addresses) ? Array.from(new Set(peer.addresses.filter(Boolean))) : [];
+      const addressLine = peerAddresses.length ? peerAddresses.map((a) => isSyncRedactionEnabled() ? redactAddress(a) : a).join(" · ") : "No addresses";
+      const addressLabel = el("div", "peer-addresses", addressLine);
+      const lastSyncAt = peerStatus.last_sync_at || peerStatus.last_sync_at_utc || "";
+      const lastPingAt = peerStatus.last_ping_at || peerStatus.last_ping_at_utc || "";
+      const meta = el(
+        "div",
+        "peer-meta",
+        [
+          lastSyncAt ? `Sync: ${formatTimestamp(lastSyncAt)}` : "Sync: never",
+          lastPingAt ? `Ping: ${formatTimestamp(lastPingAt)}` : "Ping: never"
+        ].join(" · ")
+      );
+      const identityMeta = el(
+        "div",
+        "peer-meta",
+        peer.actor_display_name ? `Assigned to ${String(peer.actor_display_name)}${peer.claimed_local_actor ? " · local actor" : ""}` : "Unassigned actor"
+      );
+      const scope = peer.project_scope || {};
+      const includeList = Array.isArray(scope.include) ? scope.include : [];
+      const excludeList = Array.isArray(scope.exclude) ? scope.exclude : [];
+      const effectiveInclude = Array.isArray(scope.effective_include) ? scope.effective_include : [];
+      const effectiveExclude = Array.isArray(scope.effective_exclude) ? scope.effective_exclude : [];
+      const inheritsGlobal = Boolean(scope.inherits_global);
+      const scopePanel = el("div", "peer-scope");
+      const identityRow = el("div", "peer-scope-summary");
+      identityRow.textContent = "Assigned actor";
+      const actorRow = el("div", "peer-actor-row");
+      const actorSelect = document.createElement("select");
+      actorSelect.className = "sync-actor-select";
+      actorSelect.setAttribute("aria-label", `Assigned actor for ${displayName}`);
+      buildActorOptions(String(peer.actor_id || "")).forEach(
+        (option) => actorSelect.appendChild(option)
+      );
+      const applyActorBtn = el("button", "settings-button", "Save actor");
+      const actorHint = el(
+        "div",
+        "peer-scope-effective",
+        assignmentNote(String(peer.actor_id || ""))
+      );
+      actorSelect.addEventListener("change", () => {
+        actorHint.textContent = assignmentNote(actorSelect.value);
+      });
+      applyActorBtn.addEventListener("click", async () => {
+        applyActorBtn.disabled = true;
+        actorSelect.disabled = true;
+        applyActorBtn.textContent = "Applying…";
+        try {
+          await assignPeerActor(peerId, actorSelect.value || null);
+          showGlobalNotice(actorSelect.value ? "Device actor updated." : "Device actor cleared.");
+          await _loadSyncData();
+        } catch (error) {
+          showGlobalNotice(friendlyError(error, "Failed to update device actor."), "warning");
+          applyActorBtn.textContent = "Retry";
+        } finally {
+          actorSelect.disabled = false;
+          applyActorBtn.disabled = false;
+          if (applyActorBtn.textContent === "Applying…")
+            applyActorBtn.textContent = "Save actor";
+        }
+      });
+      actorRow.append(actorSelect, applyActorBtn);
+      const scopeSummary = el(
+        "div",
+        "peer-scope-summary",
+        inheritsGlobal ? "Using global sync scope" : `Device override · include: ${includeList.join(", ") || "all"} · exclude: ${excludeList.join(", ") || "none"}`
+      );
+      const effectiveSummary = el(
+        "div",
+        "peer-scope-effective",
+        `Effective scope · include: ${effectiveInclude.join(", ") || "all"} · exclude: ${effectiveExclude.join(", ") || "none"}`
+      );
+      const includeEditor = createChipEditor(includeList, "Add included project", "All projects");
+      const excludeEditor = createChipEditor(excludeList, "Add excluded project", "No exclusions");
+      const editorWrap = el("div", "peer-scope-editor-wrap");
+      const scopeEditorOpen = openPeerScopeEditors.has(peerId);
+      editorWrap.hidden = !scopeEditorOpen;
+      const inputRow = el("div", "peer-scope-row");
+      inputRow.append(includeEditor.element, excludeEditor.element);
+      const scopeActions = el("div", "peer-scope-actions");
+      const saveScopeBtn = el("button", "settings-button", "Save scope");
+      const inheritBtn = el(
+        "button",
+        "settings-button",
+        "Reset to global scope"
+      );
+      saveScopeBtn.addEventListener("click", async () => {
+        saveScopeBtn.disabled = true;
+        saveScopeBtn.textContent = "Saving…";
+        try {
+          await updatePeerScope(peerId, includeEditor.values(), excludeEditor.values());
+          showGlobalNotice("Device sync scope saved.");
+          await _loadSyncData();
+        } catch (error) {
+          showGlobalNotice(friendlyError(error, "Failed to save device scope."), "warning");
+          saveScopeBtn.textContent = "Retry save";
+        } finally {
+          saveScopeBtn.disabled = false;
+          if (saveScopeBtn.textContent === "Saving…") saveScopeBtn.textContent = "Save scope";
+        }
+      });
+      inheritBtn.addEventListener("click", async () => {
+        inheritBtn.disabled = true;
+        inheritBtn.textContent = "Resetting…";
+        try {
+          await updatePeerScope(peerId, null, null, true);
+          showGlobalNotice("Device sync scope reset to global defaults.");
+          await _loadSyncData();
+        } catch (error) {
+          showGlobalNotice(friendlyError(error, "Failed to reset device scope."), "warning");
+          inheritBtn.textContent = "Retry reset";
+        } finally {
+          inheritBtn.disabled = false;
+          if (inheritBtn.textContent === "Resetting…")
+            inheritBtn.textContent = "Reset to global scope";
+        }
+      });
+      scopeActions.append(saveScopeBtn, inheritBtn);
+      editorWrap.append(inputRow, scopeActions);
+      toggleScopeBtn.textContent = scopeEditorOpen ? "Hide scope editor" : "Edit scope";
+      toggleScopeBtn.addEventListener("click", () => {
+        const nextHidden = !editorWrap.hidden;
+        editorWrap.hidden = nextHidden;
+        if (nextHidden) openPeerScopeEditors.delete(peerId);
+        else openPeerScopeEditors.add(peerId);
+        toggleScopeBtn.textContent = nextHidden ? "Edit scope" : "Hide scope editor";
+      });
+      scopePanel.append(identityRow, identityMeta, actorRow, actorHint, scopeSummary, effectiveSummary, editorWrap);
+      titleRow.append(name, actions);
+      card.append(titleRow, addressLabel, meta, scopePanel);
+      syncPeers.appendChild(card);
+    });
+  }
+  function renderLegacyDeviceClaims() {
+    const panel = document.getElementById("syncLegacyClaims");
+    const select = document.getElementById("syncLegacyDeviceSelect");
+    const button = document.getElementById("syncLegacyClaimButton");
+    const meta = document.getElementById("syncLegacyClaimsMeta");
+    if (!panel || !select || !button || !meta) return;
+    const devices = Array.isArray(state.lastSyncLegacyDevices) ? state.lastSyncLegacyDevices : [];
+    select.textContent = "";
+    meta.textContent = "";
+    if (!devices.length) {
+      panel.hidden = true;
+      return;
+    }
+    panel.hidden = false;
+    devices.forEach((device, index) => {
+      const option = document.createElement("option");
+      const deviceId = String(device.origin_device_id || "").trim();
+      if (!deviceId) return;
+      const count = Number(device.memory_count || 0);
+      const lastSeen = String(device.last_seen_at || "").trim();
+      option.value = deviceId;
+      option.textContent = count > 0 ? `${deviceId} (${count} memories)` : deviceId;
+      if (index === 0) option.selected = true;
+      select.appendChild(option);
+      if (!meta.textContent && lastSeen) {
+        meta.textContent = `Detected from older synced memories. Latest memory: ${formatTimestamp(lastSeen)}`;
+      }
+    });
+    if (!meta.textContent) {
+      meta.textContent = "Detected from older synced memories not yet attached to a current device.";
+    }
+  }
+  function initPeopleEvents(loadSyncData2) {
+    const syncActorCreateButton = document.getElementById(
+      "syncActorCreateButton"
+    );
+    const syncActorCreateInput = document.getElementById(
+      "syncActorCreateInput"
+    );
+    const syncLegacyClaimButton = document.getElementById(
+      "syncLegacyClaimButton"
+    );
+    const syncLegacyDeviceSelect = document.getElementById(
+      "syncLegacyDeviceSelect"
+    );
+    syncActorCreateButton?.addEventListener("click", async () => {
+      if (!syncActorCreateButton || !syncActorCreateInput) return;
+      const displayName = String(syncActorCreateInput.value || "").trim();
+      if (!displayName) {
+        markFieldError(syncActorCreateInput, "Enter a name for the actor.");
+        return;
+      }
+      clearFieldError(syncActorCreateInput);
+      syncActorCreateButton.disabled = true;
+      syncActorCreateInput.disabled = true;
+      syncActorCreateButton.textContent = "Creating…";
+      try {
+        await createActor(displayName);
+        showGlobalNotice("Actor created.");
+        syncActorCreateInput.value = "";
+        await loadSyncData2();
+      } catch (error) {
+        showGlobalNotice(friendlyError(error, "Failed to create actor."), "warning");
+        syncActorCreateButton.textContent = "Retry";
+        syncActorCreateButton.disabled = false;
+        syncActorCreateInput.disabled = false;
+        return;
+      }
+      syncActorCreateButton.textContent = "Create actor";
+      syncActorCreateButton.disabled = false;
+      syncActorCreateInput.disabled = false;
+    });
+    syncLegacyClaimButton?.addEventListener("click", async () => {
+      const originDeviceId = String(syncLegacyDeviceSelect?.value || "").trim();
+      if (!originDeviceId || !syncLegacyClaimButton) return;
+      if (!window.confirm(
+        `Attach old device history from ${originDeviceId} to your local actor? This updates legacy provenance for that device.`
+      ))
+        return;
+      syncLegacyClaimButton.disabled = true;
+      const originalText = syncLegacyClaimButton.textContent || "Attach device history";
+      syncLegacyClaimButton.textContent = "Attaching…";
+      try {
+        await claimLegacyDeviceIdentity(originDeviceId);
+        showGlobalNotice("Old device history attached to your local actor.");
+        await loadSyncData2();
+      } catch (error) {
+        showGlobalNotice(friendlyError(error, "Failed to attach old device history."), "warning");
+        syncLegacyClaimButton.textContent = "Retry";
+        syncLegacyClaimButton.disabled = false;
+        return;
+      }
+      syncLegacyClaimButton.textContent = originalText;
+      syncLegacyClaimButton.disabled = false;
+    });
+  }
+  async function loadSyncData() {
+    try {
+      const payload = await loadSyncStatus(true, state.currentProject || "");
+      let actorsPayload = null;
+      let actorLoadError = false;
+      try {
+        actorsPayload = await loadSyncActors();
+      } catch {
+        actorLoadError = true;
+      }
+      const statusPayload = payload.status && typeof payload.status === "object" ? payload.status : null;
+      if (statusPayload) state.lastSyncStatus = statusPayload;
+      state.lastSyncActors = Array.isArray(actorsPayload?.items) ? actorsPayload.items : [];
+      state.lastSyncPeers = payload.peers || [];
+      state.lastSyncSharingReview = payload.sharing_review || [];
+      state.lastSyncCoordinator = payload.coordinator || null;
+      state.lastSyncJoinRequests = payload.join_requests || [];
+      state.lastSyncAttempts = payload.attempts || [];
+      state.lastSyncLegacyDevices = payload.legacy_devices || [];
+      renderSyncStatus();
+      renderTeamSync();
+      renderSyncActors();
+      renderSyncSharingReview();
+      renderSyncPeers();
+      renderLegacyDeviceClaims();
+      renderSyncAttempts();
+      renderHealthOverview();
+      if (actorLoadError) {
+        const actorMeta = document.getElementById("syncActorsMeta");
+        if (actorMeta)
+          actorMeta.textContent = "Actor controls are temporarily unavailable. Peer status and sync health still loaded.";
+      }
+    } catch {
+      const syncMeta = document.getElementById("syncMeta");
+      if (syncMeta) syncMeta.textContent = "Sync unavailable";
+    }
+  }
+  async function loadPairingData() {
+    try {
+      const payload = await loadPairing();
+      state.pairingPayloadRaw = payload || null;
+      renderPairing();
+    } catch {
+      renderPairing();
+    }
+  }
+  function initSyncTab(refreshCallback) {
+    setLoadSyncData$1(loadSyncData);
+    setLoadSyncData(loadSyncData);
+    setRenderSyncPeers(renderSyncPeers);
+    initTeamSyncEvents(refreshCallback, loadSyncData);
+    initPeopleEvents(loadSyncData);
+    initDiagnosticsEvents(refreshCallback);
   }
   let settingsOpen = false;
   let previouslyFocused = null;
